app v1

+++ app/src/main/java/ru/kpfu/itis/android/weather/model/Coord.kt
@@ -4,6 +4,6 @@ import com.google.gson.annotations.SerializedName
 
 data class Coord (
 
-	@SerializedName("lon") val lon : Int,
-	@SerializedName("lat") val lat : Int
+	@SerializedName("lon") val lon : Double,
+	@SerializedName("lat") val lat : Double
 )
